refactor(stats): centralized
