--- src/app/(pages)/games/wordleverse/(game)/_components/index.js
@@ -1,3 +1,0 @@
-export * from "./Clipboard";
-export * from "./PlayAgain";
-export * from "./Stats";
