+++ src/app/(pages)/games/wordleverse/history/page.js
@@ -2,12 +2,11 @@
 
 import { useSession } from "next-auth/react";
 
-import { useHistory } from "./_lib/storage";
-
+import Calendar from "./_components/Calendar";
 import Layout from "./_components/Layout";
 import SignInPrompt from "./_components/SignInPrompt";
 import Stats from "./_components/Stats";
-import Calendar from "./_components/Calendar";
+import { useHistory } from "./_lib/storage";
 
 export default function HistoryPage() {
   const { data: session } = useSession();
@@ -24,7 +23,7 @@ export default function HistoryPage() {
   return (
     <Layout>
       {!session && <SignInPrompt />}
-      <Stats history={history} />
+      <Stats history={history} historyPage={true} showHistoryLink={false} />
       <Calendar availableDates={availableDates} games={history.games} />
     </Layout>
   );
